+++ server/src/compiler_analyzer/hoist.ts
@@ -27,7 +27,7 @@ import {complementHintForScope} from "./complementHint";
 import {SymbolFunction, SymbolType, SymbolVariable} from "./symbolObject";
 import {findSymbolWithParent} from "./symbolUtils";
 import {ResolvedType} from "./resolvedType";
-import {getGlobalSettings} from "../code/settings";
+import {getGlobalSettings} from "../core/settings";
 import {builtinSetterValueToken, builtinThisToken, tryGetBuiltInType} from "./symbolBuiltin";
 import {TokenIdentifier, TokenObject} from "../compiler_tokenizer/tokenObject";
 import {getIdentifierInNodeType} from "../compiler_parser/nodesUtils";
