+++ server/src/core/diagnostic.ts
