+++ server/src/core/highlight.ts
