+++ server/src/core/settings.ts
